+++ podcastindex-sdk/src/commonMain/kotlin/com/mr3y/podcastindex/model/TrendingFeed.kt
@@ -15,7 +15,7 @@ public class TrendingResult(
 @Serializable
 @Poko
 public class TrendingFeed(
-    @SerialName(value = "id") public val id: Int,
+    @SerialName(value = "id") public val id: Long,
     @SerialName(value = "url") public val url: String,
     @SerialName(value = "title") public val title: String,
     @SerialName(value = "description") public val description: String,
@@ -23,7 +23,7 @@ public class TrendingFeed(
     @SerialName(value = "image") public val image: String,
     @SerialName(value = "artwork") public val artwork: String,
     @SerialName(value = "newestItemPublishTime") @Serializable(InstantSerializer::class) public val newestItemPublishTime: Instant,
-    @SerialName(value = "itunesId") public val itunesId: Int? = null,
+    @SerialName(value = "itunesId") public val itunesId: Long? = null,
     @SerialName(value = "trendScore") public val trendScore: Int,
     @SerialName(value = "language") public val language: String,
 )
